backend/requirements.txt
fastapi
uvicorn
langgraph
pydantic
python-multipart
PyPDF2
requests
python-dotenv